Integration of viewer-count-cdk-stack
